@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^20.1.3",
-    "@angular/cdk": "^20.1.3",
-    "@angular/common": "^20.1.3",
-    "@angular/compiler": "^20.1.3",
-    "@angular/core": "^20.1.3",
-    "@angular/forms": "^20.1.3",
-    "@angular/platform-browser": "^20.1.3",
-    "@angular/platform-server": "^20.1.3",
-    "@angular/router": "^20.1.3",
-    "@angular/ssr": "^20.1.3",
+    "@angular/animations": "^20.3.6",
+    "@angular/cdk": "^20.2.9",
+    "@angular/common": "^20.3.6",
+    "@angular/compiler": "^20.3.6",
+    "@angular/core": "^20.3.6",
+    "@angular/forms": "^20.3.6",
+    "@angular/platform-browser": "^20.3.6",
+    "@angular/platform-server": "^20.3.6",
+    "@angular/router": "^20.3.6",
+    "@angular/ssr": "^20.3.6",
     "@fortawesome/angular-fontawesome": "^3.0.0",
     "@fortawesome/free-regular-svg-icons": "^7.1.0",
     "@fortawesome/free-solid-svg-icons": "^7.1.0",
@@ -35,9 +35,9 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular/build": "^20.1.3",
-    "@angular/cli": "^20.1.3",
-    "@angular/compiler-cli": "^20.1.3",
+    "@angular/build": "^20.3.6",
+    "@angular/cli": "^20.3.6",
+    "@angular/compiler-cli": "^20.3.6",
     "@types/express": "^5.0.1",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^20.17.19",
